# opsplatform-video-manager-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CDNProvider:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateDomainRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  models.CreateStreamRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      provider_id:
        description: 'Optional: NULL means match all providers'
        type: integer
    required:
    - code
    - name
    type: object
...
  models.Domain:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Stream:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.CDNProvider'
      provider_id:
        description: NULL means match all providers
        type: integer
      updated_at:
        type: string
    type: object
...
  models.UpdateDomainRequest:
    properties:
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  models.UpdateStreamRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      provider_id:
        description: 'Optional: NULL means match all providers'
        type: integer
    required:
    - code
    - name
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Video Manager API Documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Video Manager API
  version: "1.0"
paths:
...
  /api/domains:
    get:
      description: Retrieve all domains
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Domain'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all domains
      tags:
      - domains
    post:
      consumes:
      - application/json
      description: Create a new domain with the provided information
      parameters:
      - description: Domain information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Domain'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new domain
      tags:
      - domains
  /api/domains/{id}:
    delete:
      description: Delete a domain by its ID. This will also delete associated video
        stream endpoints.
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a domain
      tags:
      - domains
    get:
      description: Retrieve a domain by its ID
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Domain'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get domain by ID
      tags:
      - domains
    put:
      consumes:
      - application/json
      description: Update an existing domain by its ID
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated domain information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Domain'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a domain
      tags:
      - domains
...
  /api/stream-regions:
    get:
      description: Retrieve all streams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Stream'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all streams
      tags:
      - stream-regions
    post:
      consumes:
      - application/json
      description: Create a new stream with the provided information
      parameters:
      - description: Stream information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Stream'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new stream
      tags:
      - stream-regions
  /api/stream-regions/{id}:
    delete:
      description: Delete a stream by its ID. This will also delete associated stream
        paths and video stream endpoints.
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a stream
      tags:
      - stream-regions
    get:
      description: Retrieve a stream by its ID
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Stream'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get stream by ID
      tags:
      - stream-regions
    put:
      consumes:
      - application/json
      description: Update an existing stream by its ID
      parameters:
      - description: Stream ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated stream information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Stream'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a stream
      tags:
      - stream-regions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"